refactor: Update README.md with additional step for cleaning local storage
This code change updates the README.md file to include an additional step for cleaning local storage in the browser. This step is necessary to ensure that any previous data stored in the local storage does not interfere with the functionality of the application. By including this step, users will have a clear instruction on how to clean their local storage before running the application.