refactor: Add default value for isPaid field in Order model
This code change adds a default value of `false` for the `isPaid` field in the Order model. By setting a default value, it ensures that the `isPaid` field is automatically set to `false` when a new order is created. This enhancement improves the consistency and reliability of the order management system.