feat: Add clearCart method to cart-store
This code change adds a `clearCart` method to the `cart-store` in order to clear the cart. The method sets the `cart` array to an empty array, effectively removing all items from the cart. This enhancement improves the functionality of the application by providing a convenient way for users to clear their cart when needed.