feat: Update PayPal payment status in database, Error here
This code change adds functionality to update the payment status in the database when a PayPal payment is completed. It uses the `prisma` library to update the `isPaid` field of the corresponding order to `true` and sets the `paidAt` field to the current date and time. If an error occurs during the update, an error message is returned. This enhancement improves the accuracy of payment records in the application.