feat: Update PayPal payment status in database
This code change updates the payment status in the database when a PayPal payment is completed. It uses the `prisma` library to update the `isPaid` field of the corresponding order to `true` and sets the `paidAt` field to the current date and time. This enhancement improves the accuracy of payment records in the application.