feat: Add changeUserRole function
This code change adds the `changeUserRole` function, which allows an admin user to change the role of another user. The function requires authentication and only allows admins to access it. It updates the role of the specified user in the database and triggers a revalidation of the `/admin/users` path. This addition enhances the application by providing a way for admins to manage user roles.