feat: Refactor getProductbySlug to include product image IDs
This code change refactors the getProductbySlug function to include the IDs of the product images in the returned data. Previously, only the image URLs were included. This modification enhances the functionality by providing access to the image IDs, which can be used for various purposes such as image deletion or updating.