feat: Update ProductForm to save product data
This code change updates the ProductForm component to include the functionality to save product data. When the form is submitted, the data is collected and converted into a FormData object. This FormData object is then passed to the createUpdateProduct function, which handles the API request to create or update the product. This enhancement improves the functionality of the application by allowing users to save changes made to the product form.