feat: Add createUpdateProduct function for saving product data
This code change adds the createUpdateProduct function, which is responsible for saving product data. The function takes in a FormData object and validates the data using a schema. If the data is valid, the function returns a success message. If the data is invalid, an error message is returned. This enhancement improves the functionality of the application by allowing users to create or update products with validated data.