feat: Replace Image component with ProductImage in ProductsInCart
This code change replaces the usage of the Image component with the newly added ProductImage component in the ProductsInCart file. The ProductImage component is responsible for displaying product images and provides better performance by utilizing lazy loading. This enhancement improves the visual presentation of product images in the cart and optimizes the overall performance of the application.